requirements.txt
tensorflow
keras==2.0.9
pillow
h5py
